move implementions of util functions from ds3.c to ds3_utils.c, to match definitions in ds3_utils.h
